Сдача проектной работы

+++ src/scripts/card.js
@@ -1,16 +1,27 @@
-const cardTemplate = document.querySelector("#card-template").content;
-const placesList = document.querySelector(".places__list");
+import { deleteCardAPI, likeCardAPI, unlikeCardAPI } from "./api";
+import { closeModal } from "./modal";
 
-function createCard(item, { deleteCard, likeCard, openImageModal }) {
+const cardTemplate = document.querySelector("#card-template").content;
+const popupDeleteCard = document.querySelector(".popup__delete-card");
+
+function createCard(
+  item,
+  { deleteCard, openImageModal, openDeleteModal },
+  userInfo
+) {
   const cardElement = cardTemplate.querySelector(".card").cloneNode(true);
   const cardImage = cardElement.querySelector(".card__image");
   const cardTitle = cardElement.querySelector(".card__title");
   const deleteButton = cardElement.querySelector(".card__delete-button");
   const likeButton = cardElement.querySelector(".card__like-button");
+  const likeCounter = cardElement.querySelector(".like__counter");
+  const buttonConfirm = document.querySelector(".button__confirm");
 
   cardImage.src = item.link;
   cardImage.alt = item.name;
   cardTitle.textContent = item.name;
+
+  likeCounter.textContent = item.likes.length;
 
   cardImage.addEventListener("click", () => {
     const imageSrc = item.link;
@@ -18,14 +29,43 @@ function createCard(item, { deleteCard, likeCard, openImageModal }) {
     openImageModal(imageSrc, imageCaption);
   });
 
-  deleteButton.addEventListener("click", () => {
-    deleteCard(cardElement);
-  });
+  if (userInfo._id === item.owner._id) {
+    deleteButton.addEventListener("click", () => {
+      openDeleteModal();
+
+      buttonConfirm.addEventListener("click", () => {
+        deleteCardAPI(item._id).then(() => {
+          deleteCard(cardElement);
+          closeModal(popupDeleteCard);
+        });
+      });
+    });
+  } else {
+    deleteButton.style.display = "none";
+  }
 
   likeButton.addEventListener("click", () => {
-    likeCard(likeButton);
+    const isSimillar = item.likes.find((el) => userInfo._id === el._id);
+
+    if (isSimillar) {
+      unlikeCardAPI(item._id).then((res) => {
+        item.likes = res.likes;
+        likeCounter.textContent = res.likes.length;
+      });
+      likeButton.classList.remove("card__like-button_is-active");
+    } else {
+      likeCardAPI(item._id).then((res) => {
+        item.likes = res.likes;
+        likeCounter.textContent = res.likes.length;
+      });
+      likeButton.classList.add("card__like-button_is-active");
+    }
   });
 
+  const exist = item.likes.find((el) => userInfo._id === el._id);
+  if (exist) {
+    likeButton.classList.add("card__like-button_is-active");
+  }
   return cardElement;
 }
 
@@ -33,8 +73,4 @@ function deleteCard(card) {
   card.remove();
 }
 
-function likeCard(likeButton) {
-  likeButton.classList.toggle("card__like-button_is-active");
-}
-
-export { createCard, deleteCard, likeCard, placesList };
+export { createCard, deleteCard };
